refactor: implement gromosXX Algorithm/AlgorithmSequence pattern
- Add Algorithm trait and AlgorithmSequence in gromos-core with 7 unit tests
- Implement algorithm wrappers: Forcefield, LeapFrogVelocity, LeapFrogPosition,
  TemperatureCalculation, EnergyCalculation
- TemperatureCalculation uses gromosXX averaged kinetic energy formula:
  E_kin = 0.5 * sum(m_i * (|v_new|^2 + |v_old|^2) / 2)
- Refactor md.rs to use AlgorithmSequence instead of manual force+integrate loop
- Energy convention follows gromosXX: write to current(), exchange_state moves
  to old(), read from old() for output
- Energy output uses full f64 scientific notation for exact comparison
- All 10 steps of pair_lj match gromosXX reference to ~1e-11 precision
- Fix minor clippy warnings in gamd, remd, integrator, pairlist

+++ crates/gromos-core/src/pairlist.rs
@@ -556,7 +556,7 @@ mod tests {
 
         // Place atoms randomly in box
         for (i, pos) in conf.current_mut().pos.iter_mut().enumerate() {
-            *pos = Vec3::new((i % 5.0), ((i / 5) % 5.0), 0.0);
+            *pos = Vec3::new((i % 5) as f64, ((i / 5) % 5) as f64, 0.0);
         }
 
         let mut pairlist = PairlistContainer::new(1.5, 2.5, 0.2);
